backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.11.0
pymongo==4.6.1
psycopg2-binary==2.9.9
torch==2.1.2
transformers==4.41.0
numpy==1.26.4
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.1.0
tqdm==4.66.1
